test-requirements.txt
django
gunicorn
selenium
fabric3
Pillow
factory_boy
